refactor(utilities): split grid helpers
